Fix packet_delay_in_nano_seconds_

+++ apps/examples/dvb/packet_sender.cpp
@@ -17,7 +17,7 @@ PacketSender::PacketSender(srslog::basic_logger&   logger_,
                            dvb_tx_sim_transceiver& transceiver_,
                            PacketQueue&            queue_,
                            kpi_counter&            tx_bytes_,
-                           uint16_t                packet_delay_in_nano_seconds_) :
+                           unsigned                packet_delay_in_nano_seconds_) :
   logger(logger_),
   executor(executor_),
   packets(queue_),
