+++ apps/examples/dvb/packet_sender.h
@@ -17,7 +17,7 @@ class PacketSender
   PacketQueue&            packets;
   dvb_tx_sim_transceiver& transceiver;
   kpi_counter&            tx_bytes;
-  uint16_t                packet_delay_in_nano_seconds;
+  unsigned                packet_delay_in_nano_seconds;
 
   void send_loop();
 
@@ -27,7 +27,7 @@ public:
                dvb_tx_sim_transceiver& transceiver_,
                PacketQueue&            queue_,
                kpi_counter&            tx_bytes_,
-               uint16_t                packet_delay_in_nano_seconds_);
+               unsigned                packet_delay_in_nano_seconds_);
   void start();
 };
 } // namespace srsran
